Revert "Merge pull request #132 from navikt/menu-style"
This reverts commit 4a5ce57386874efe4fadab249a3afc28e0114069, reversing
changes made to 6d90aa3acd5fe0a8368d9dc87c2b755a81e43e5b.
