+++ src/components/PaginationRow.tsx
@@ -1,12 +1,10 @@
-import React, { ReactNode } from 'react';
+import React from 'react';
 import {
     useEffect,
     useState,
 } from 'react';
 import styled from 'styled-components';
 import ChevronKnapp from './ChevronKnapp';
-import themes from '../styles/themes';
-import Clickable from './toolbar/Clickable';
 
 const PaginationItems = styled.div`
   display: flex;
@@ -15,20 +13,18 @@ const PaginationItems = styled.div`
   justify-content: space-between;
 `;
 
+const PaginationText = styled.div`
+  display: inline-flex;
+  justify-items: center;
+  white-space: nowrap;
+`;
+
 interface PaginationProps {
   startPage?: number;
   numberOfItems: number;
   maxNumberPerPage: number;
   onPageChange(start: number, end: number, pageNumber: number): void;
 }
-
-const Element = styled.div`
-  display: flex;
-  justify-content: center;
-  border-left: 2px solid ${themes.color.navGra40};
-  width: 2.5em;
-`;
-
 const PaginationRow = ({
   numberOfItems,
   startPage = 0,
@@ -73,22 +69,16 @@ const PaginationRow = ({
     };
   };
 
+  const getCurrentPageAsNormalizedText = () => `${currentPage + 1}`;
+  const getLastPageAsNormalizedText = () => `${getNumberOfPages() + 1}`;
+
   return (
     <PaginationItems>
-      <Element>
-        <ChevronKnapp disabled={currentPage === 0} type="venstre" onClick={onPreviousClick} />
-      </Element>
-      {currentPage !== getNumberOfPages() &&
-        <Element>
-          <Clickable index={currentPage} setPage={setCurrentPage} currentPage={currentPage} />
-        </Element>
-      }
-      <Element>
-        <Clickable index={getNumberOfPages()} setPage={setCurrentPage} currentPage={currentPage} />
-      </Element>
-      <Element>
-        <ChevronKnapp disabled={currentPage === getNumberOfPages()} type="høyre" onClick={onNextClick} />
-      </Element>
+        <ChevronKnapp visible={currentPage !== 0} type="venstre" onClick={onPreviousClick} />
+        <PaginationText>
+            <p>Side <strong>{getCurrentPageAsNormalizedText()}</strong> av <strong>{getLastPageAsNormalizedText()}</strong></p>
+        </PaginationText>
+        <ChevronKnapp visible={currentPage !== getNumberOfPages()} type="høyre" onClick={onNextClick} />
     </PaginationItems>
   );
 };
