+++ src/components/toolbar/PaginationContainer.tsx
@@ -9,7 +9,6 @@ import themes from '../../styles/themes';
 
 const TogglePagination = styled.p`
   cursor: pointer;
-  margin-right: .7em;
   color: ${themes.color.navBla};
   :hover {
     border-bottom: 1px solid ${themes.color.navGra40};
